scanfusion-backend/requirements.txt
fastapi
uvicorn
torch
monai
nibabel
numpy
matplotlib
pillow
opencv-python-headless